Add support for --java argument to CDH topology
This commit introduces an optional `volumes_from` argument to `Node`, which is then used by the
CDH topology to support selection of which Java version to use for a cluster deployment at runtime.
For this functionality to work, a Docker image tagged with the same repository name as the CDH
node images must be built using the Dockerfile in the topology's `jdk` folder. This image's
tag must be the directory name of the Dockerfile's `JDK_DIRECTORY` build arg with `cdh_` prepended
(e.g. `cdh_jdk1.8.0_121`).